requirements.txt
streamlit
nltk
pymupdf
sentence-transformers
faiss-cpu
langchain_groq
python-dotenv
